+++ peer/set.go
@@ -1,47 +1,45 @@
-package peerset
+package peer
 
 import (
 	"sync"
-
-	"github.com/libp2p/go-libp2p-core/peer"
 )
 
 // PeerSet is a threadsafe set of peers
-type PeerSet struct {
+type Set struct {
 	lk sync.RWMutex
-	ps map[peer.ID]struct{}
+	ps map[ID]struct{}
 
 	size int
 }
 
-func New() *PeerSet {
-	ps := new(PeerSet)
-	ps.ps = make(map[peer.ID]struct{})
+func NewSet() *Set {
+	ps := new(Set)
+	ps.ps = make(map[ID]struct{})
 	ps.size = -1
 	return ps
 }
 
-func NewLimited(size int) *PeerSet {
-	ps := new(PeerSet)
-	ps.ps = make(map[peer.ID]struct{})
+func NewLimitedSet(size int) *Set {
+	ps := new(Set)
+	ps.ps = make(map[ID]struct{})
 	ps.size = size
 	return ps
 }
 
-func (ps *PeerSet) Add(p peer.ID) {
+func (ps *Set) Add(p ID) {
 	ps.lk.Lock()
 	ps.ps[p] = struct{}{}
 	ps.lk.Unlock()
 }
 
-func (ps *PeerSet) Contains(p peer.ID) bool {
+func (ps *Set) Contains(p ID) bool {
 	ps.lk.RLock()
 	_, ok := ps.ps[p]
 	ps.lk.RUnlock()
 	return ok
 }
 
-func (ps *PeerSet) Size() int {
+func (ps *Set) Size() int {
 	ps.lk.RLock()
 	defer ps.lk.RUnlock()
 	return len(ps.ps)
@@ -51,7 +49,7 @@ func (ps *PeerSet) Size() int {
 // This operation can fail for one of two reasons:
 // 1) The given peer is already in the set
 // 2) The number of peers in the set is equal to size
-func (ps *PeerSet) TryAdd(p peer.ID) bool {
+func (ps *Set) TryAdd(p ID) bool {
 	var success bool
 	ps.lk.Lock()
 	if _, ok := ps.ps[p]; !ok && (len(ps.ps) < ps.size || ps.size == -1) {
@@ -62,9 +60,9 @@ func (ps *PeerSet) TryAdd(p peer.ID) bool {
 	return success
 }
 
-func (ps *PeerSet) Peers() []peer.ID {
+func (ps *Set) Peers() []ID {
 	ps.lk.Lock()
-	out := make([]peer.ID, 0, len(ps.ps))
+	out := make([]ID, 0, len(ps.ps))
 	for p, _ := range ps.ps {
 		out = append(out, p)
 	}
