+++ routing/query.go
@@ -1,4 +1,4 @@
-package notifications
+package routing
 
 import (
 	"context"
